#9: Refactoring of the price engine module in different submodules, fixed mypy issues
