+++ jaxfin/price_engine/math/bs_common.py
@@ -1,11 +1,11 @@
 """
-Common functions for the price_engine submodules
+Common math functions related to the Black Scholes model for the price_engine submodules
 """
 
 import jax
 import jax.numpy as jnp
 
-from .math import cum_normal, d1
+from .norm import cum_normal
 
 
 def _compute_d1_d2(spots, strikes, expires, vols, discount_rates):
@@ -24,6 +24,26 @@ def _compute_d1_d2(spots, strikes, expires, vols, discount_rates):
     _d1 = d1(spots, strikes, vols, expires, discount_rates)
 
     return [_d1, _d1 - vol_sqrt_t]
+
+
+@jax.jit
+def d1(spots, strikes, vols, expires, discount_rates):
+    """
+    Calculate the d1 term in the Black-Scholes formula
+
+    :param spots: Current spot price of the underlying
+    :param strikes: Strike price of the option
+    :param vols: Volatility of the underlying
+    :param expires: Time to expiration of the option
+    :param discount_rates: Risk-free rate
+    :return: d1 term
+    """
+    vol_sqrt_t = vols * jnp.sqrt(expires)
+
+    return jnp.divide(
+        (jnp.log(spots / strikes) + (discount_rates + (vols**2 / 2)) * expires),
+        vol_sqrt_t,
+    )
 
 
 @jax.jit
