+++ tests/price_engine/test_black.py
@@ -3,6 +3,7 @@ import jax.numpy as jnp
 import pytest
 
 from jaxfin.price_engine.black import future_option_price, future_option_delta, future_option_gamma
+from jaxfin.price_engine.utils.vect import get_vfunction
 
 TOL = 1e-3
 DTYPE = jnp.float32
@@ -75,7 +76,7 @@ def test_foption_batch_mixed_disc():
 
 class TestDelta:
 
-    def test_delta_bs_float(self):
+    def test_delta_float(self):
         spot = jnp.array(100, dtype=DTYPE)
         strike = jnp.array(120, dtype=DTYPE)
         expire = jnp.array(1, dtype=DTYPE)
@@ -90,19 +91,21 @@ class TestDelta:
         assert jnp.isclose(call_delta, expected_delta, atol=TOL)
         assert jnp.isclose(put_delta, expected_put_delta, atol=TOL)
 
-    def test_delta_bs(self):
+    def test_delta(self):
         spots = jnp.array([100, 100, 100, 80, 170], dtype=DTYPE)
         strikes = jnp.array([120, 110, 120, 150, 160], dtype=DTYPE)
         expires = jnp.array([1, 1, 1, 0.5, 0.25], dtype=DTYPE)
         vols = jnp.array([0.3, 0.3, 0.2, 0.5, 0.15], dtype=DTYPE)
         rates = jnp.array([0.0, 0.0, 0.05, 0.02, 0.01], dtype=DTYPE)
         dividends = jnp.array([0.0, 0.0, 0.0, 0.0, 0.0], dtype=DTYPE)
+        call_flags = jnp.array([True, True, True, True, True], dtype=jnp.bool_)
         put_flags = jnp.array([False, False, False, False, False], dtype=jnp.bool_)
-        e_call_deltas = jnp.array([0.32357, 0.43341, 0.30971, 0.05894, 0.81452], dtype=DTYPE)
+        e_call_deltas = jnp.array([0.32357, 0.43340, 0.28024, 0.05777, 0.81046], dtype=DTYPE)
         e_put_deltas = jnp.array([-0.67643, -0.56659, -0.71975, -0.94222, -0.18953], dtype=DTYPE)
+        vmap_delta = get_vfunction(future_option_delta, spots, strikes, expires, vols, rates, dividends, are_calls=put_flags)
 
-        call_deltas = future_option_delta(spots, strikes, expires, vols, rates, dividends)
-        put_deltas = future_option_delta(spots, strikes, expires, vols, rates, dividends, are_calls=put_flags)
+        call_deltas = vmap_delta(spots, strikes, expires, vols, rates, dividends, call_flags)
+        put_deltas = vmap_delta(spots, strikes, expires, vols, rates, dividends, put_flags)
 
         assert jnp.allclose(call_deltas, e_call_deltas, atol=TOL)
         assert jnp.allclose(put_deltas, e_put_deltas, atol=TOL)
@@ -118,14 +121,14 @@ class TestDelta:
 class TestGamma:
 
     def test_gamma_bs(self, spot, strike, expire, vol, rate, expected_gamma):
-        spot = jnp.array([spot], dtype=DTYPE)
-        strike = jnp.array([strike], dtype=DTYPE)
-        expire = jnp.array([expire], dtype=DTYPE)
-        vol = jnp.array([vol], dtype=DTYPE)
-        rate = jnp.array([rate], dtype=DTYPE)
-        dividends = jnp.array([0.0], dtype=DTYPE)
-        put_flag = jnp.array([False], dtype=jnp.bool_)
-        expected_gamma = jnp.array([expected_gamma], dtype=DTYPE)
+        spot = jnp.array(spot, dtype=DTYPE)
+        strike = jnp.array(strike, dtype=DTYPE)
+        expire = jnp.array(expire, dtype=DTYPE)
+        vol = jnp.array(vol, dtype=DTYPE)
+        rate = jnp.array(rate, dtype=DTYPE)
+        dividends = jnp.array(0.0, dtype=DTYPE)
+        put_flag = jnp.array(False, dtype=jnp.bool_)
+        expected_gamma = jnp.array(expected_gamma, dtype=DTYPE)
         call_gamma = future_option_gamma(spot, strike, expire, vol, rate, dividends)
         put_gamma = future_option_gamma(spot, strike, expire, vol, rate, dividends, are_calls=put_flag)
